requirements.txt
psycopg2-binary==2.9.10
pydantic==2.10.4
pandas==2.2.3
openpyxl==3.1.5
PyYAML==6.0.2
python-dotenv==1.0.1